--- src/main/java/org/lab41/web/controller/UserController.java
@@ -1,43 +1,0 @@
-/**
- * Copyright 2013 In-Q-Tel/Lab41
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.lab41.web.controller;
-
-import org.codehaus.jackson.map.ObjectMapper;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Controller
-public class UserController {
-
-    @RequestMapping(value = "/api/user", method = RequestMethod.GET)
-    public @ResponseBody Map<String,Object> userInformation() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        Map<String, Object> response = new HashMap<String, Object>();
-        response.put("name", authentication.getName());
-        response.put("authorities", authentication.getAuthorities());
-
-        return response;
-    }
-}
