+++ src/test/java/org/lab41/dendrite/faunus/AdjacencyListInputFormatTest.java
@@ -1,0 +1,13 @@
+package org.lab41.dendrite.faunus;
+
+import static org.junit.Assert.assertTrue;
+
+public class AdjacencyListInputFormatTest {
+
+    /*
+    @Test
+    public void testInGremlinImports() {
+        assertTrue(Imports.getImports().contains(AdjacencyFileInputFormat.class.getPackage().getName() + ".*"));
+    }
+    */
+}
