+++ src/test/java/org/lab41/dendrite/faunus/AdjacencyListOutputFormatTest.java
@@ -1,0 +1,13 @@
+package org.lab41.dendrite.faunus;
+
+import static org.junit.Assert.assertTrue;
+
+public class AdjacencyListOutputFormatTest {
+
+    /*
+    @Test
+    public void testInGremlinImports() {
+        assertTrue(Imports.getImports().contains(AdjacencyFileOutputFormat.class.getPackage().getName() + ".*"));
+    }
+    */
+}
