--- src/test/java/org/lab41/faunus/AdjacencyListOutputFormatTest.java
@@ -1,16 +1,0 @@
-package org.lab41.faunus;
-
-import com.thinkaurelius.faunus.tinkerpop.gremlin.Imports;
-import org.junit.Test;
-
-import static org.junit.Assert.assertTrue;
-
-public class AdjacencyListOutputFormatTest {
-
-    /*
-    @Test
-    public void testInGremlinImports() {
-        assertTrue(Imports.getImports().contains(AdjacencyFileOutputFormat.class.getPackage().getName() + ".*"));
-    }
-    */
-}
