+++ src/test/resources/org/lab41/dendrite/faunus/graph-of-the-gods.json
